Move menu related code to a file
