+++ src/graphics.rs
@@ -7,3 +7,4 @@ pub mod rectangle;
 pub mod tile;
 pub mod color;
 pub mod font;
+pub mod menu;
